Extract event action orchestrators into lib/event-actions.server.ts
Lifts the create/update/rsvp action bodies out of dashboard.events.tsx into
reusable helpers (runCreateEventAction, runUpdateEventAction, runRsvpEventAction)
so dashboard._index.tsx can dispatch the same flows without duplicating ~150
lines of email-batch staging and activity logging.

The events route now delegates entirely to these helpers; behavior is preserved
end-to-end (all 617 tests still pass).

Prep for #145 (events migration onto the dashboard).
